some updates...

+++ src/Controller/OutingController.php
@@ -104,7 +104,7 @@ class OutingController extends AbstractController
 
         $outingForm->handleRequest($request);
 
-        if ($outingForm->isSubmitted() && $outingForm->isValid()) {
+        if ($outingForm->isSubmitted() && $outingForm->isValid() && $this->getUser()->getId() === $outing->getPlanner()->getId()) {
 
             $outingRepository->save($outing, true);
 
@@ -130,7 +130,7 @@ class OutingController extends AbstractController
 {
     $outing = $outingRepository->find($id);
 
-    if ($outing) {
+    if ($outing && $this->getUser()->getId() === $outing->getPlanner()->getId()) {
         $outingRepository->remove($outing);
         $this->addFlash("warning", "La sortie a été supprimée, cette action est irréversible");
     } else {
@@ -149,7 +149,7 @@ class OutingController extends AbstractController
     {
         $outing = $outingRepository->find($id);
 
-        if ($outing) {
+        if ($outing && $this->getUser()->getId() === $outing->getPlanner()->getId()) {
             $outing->setStatus($statusRepository->find(5));
             $this->addFlash("warning", "Votre sortie a été annulée");
         } else {
